HW1/requirements.txt
cycler==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.0
numpy==1.20.1
pandas==1.2.3
Pillow==8.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
six==1.15.0
threadpoolctl==2.1.0
tqdm==4.59.0
